// src/reportes-bases/interfaces/valorizado.interface.ts
export interface valorizadoBod{
  CODIGO_BODEGA: number;      
  NOMBRE_BODEGA: number;      
  CODIGO_SUCURSAL: number;    
  PRODUCT0: string;           
  DESCRIPCION_CORTA: string;  
  DESCRIPCION_LARGA: string;  
  CODIGO_LOTE: number;        
  DESCRIPCION_LOTE: string;   
  FECHA_VENCIMIENTO: Date;    
  CANTIDAD_DISPONIBLE: number;
  CANTIDAD_RESERVADA: number; 
  VENCE: string;               
  DESCRIPCION_TIPO: string;   
  NOMBRE_MARCA: string;       
  COSTO_PROMEDIO: number;     
}